lib/agent-platform-core/pyproject.toml
[project]
name = "agent-platform-core"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_platform"]
